pyproject.toml
[tool.poetry]
name = "pytest-prettifier"
version = "0.0.0"
description = "Pytest fixture to prettify test results."
authors = ["Marc Tardif"]
readme = "README.rst"
repository = "https://github.com/cr3/pytest-prettifier"
packages = [
    { include = "pytest_prettifier" },
]

[tool.poetry.dependencies]
python = "^3.9"
attrs = "^24.3.0"
python-dateutil = "^2.9.0.post0"

[tool.poetry.group.test.dependencies]
coverage = "^7.2.3"
pytest = "^7.2.2"

[tool.poetry.group.check]
optional = true

[tool.poetry.group.check.dependencies]
ruff = "^0.0.265"
black = "^23.3.0"
pre-commit = "^3.3.1"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = "^6.1.3"
sphinxcontrib-log-cabinet = "^1.0.1"
sphinx-rtd-theme = "^1.2.0"

[tool.poetry.plugins."pytest_prettifier"]
bytes = "pytest_prettifier.prettifier:bytes_prettifier"
datetime = "pytest_prettifier.prettifier:datetime_prettifier"
dict = "pytest_prettifier.prettifier:dict_prettifier"
exception = "pytest_prettifier.prettifier:exception_prettifier"
list = "pytest_prettifier.prettifier:list_prettifier"
mock = "pytest_prettifier.prettifier:mock_prettifier"
object = "pytest_prettifier.prettifier:object_prettifier"
re = "pytest_prettifier.prettifier:re_prettifier"
set = "pytest_prettifier.prettifier:set_prettifier"
str = "pytest_prettifier.prettifier:str_prettifier"
timedelta = "pytest_prettifier.prettifier:timedelta_prettifier"
tuple = "pytest_prettifier.prettifier:tuple_prettifier"
type = "pytest_prettifier.prettifier:type_prettifier"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ["py39"]
preview = true

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
    # Create your own exception
    "TRY002",
    # Avoid specifying long messages outside the exception class
    "TRY003",
]

[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

# Pytest options:
#   https://docs.pytest.org/en/6.2.x/reference.html#ini-options-ref
[tool.pytest.ini_options]
addopts = [
    "--doctest-modules",
    "--doctest-glob=*.rst",
]
testpaths = [
    "pytest_prettifier",
    "docs",
    "tests",
]

# Coverage options:
#   https://coverage.readthedocs.io/en/latest/config.html
[tool.coverage.paths]
source = [
    "pytest_prettifier",
    "*/*/site-packages",
]

[tool.coverage.report]
fail_under = 90
show_missing = true

[tool.coverage.run]
branch = true
parallel = true
source = [
    "pytest_prettifier",
]
